remove some unused packages + changed some codes
