+++ client/package.json
@@ -12,14 +12,12 @@
   "dependencies": {
     "@react-native-community/async-storage": "^1.9.0",
     "@react-native-community/masked-view": "^0.1.9",
-    "@react-native-community/voice": "github:react-native-community/voice",
     "@react-navigation/native": "^5.1.6",
     "@react-navigation/stack": "^5.2.11",
     "prop-types": "^15.7.2",
     "react": "16.9.0",
     "react-native": "^0.62.2",
     "react-native-gesture-handler": "^1.6.1",
-    "react-native-modal": "^11.5.6",
     "react-native-reanimated": "^1.7.0",
     "react-native-safe-area-context": "^0.7.3",
     "react-native-screens": "^2.4.0",
